bcm4329 trivial tethering bug
This prevents the SoftapController from starting wifi tether, because the
driver returns 1 instead of 0 for a Start ioctl. (If you trace the code, the
return value of 1 was hard coded. I suppose we could also just hardcode that
to 0, or change the function to be void. whatever.)

Tested with the current 3.2 kernel and the 4.217.95.0 firmware. I think it
will also work with the 4.217.74.0 firmware, haven't double-checked.